feat:Apply modal close onclick outside(#10)
